[추가] 공용 에러 핸들러 추가
