+++ src/api/shop/ShopApi.ts
@@ -1,10 +1,6 @@
-import instance from "../axios";
-import { AxiosError } from "axios";
+import instance from "@/api/axios";
+import { handleApiError } from "../error/ErrorHandler";
 import { ShopRequest, ShopResponse } from "@/types/shop";
-
-interface ErrorMessage {
-  message: string;
-}
 
 // POST '/shops' - 가게 등록
 export const postShop = async (body: ShopRequest): Promise<ShopResponse> => {
@@ -12,13 +8,7 @@ export const postShop = async (body: ShopRequest): Promise<ShopResponse> => {
     const response = await instance.post<ShopResponse>("/shops", body);
     return response.data;
   } catch (error) {
-    const axiosError = error as AxiosError<ErrorMessage>; //  에러 타입 명시
-
-    if (axiosError.response) {
-      throw new Error(axiosError.response.data.message);
-    } else {
-      throw new Error("서버를 연결할 수 없습니다. 인터넷 연결을 확인해주세요.");
-    }
+    return handleApiError(error);
   }
 };
 
@@ -28,13 +18,7 @@ export const getShop = async (shopId: string): Promise<ShopResponse> => {
     const response = await instance.get<ShopResponse>(`/shops/${shopId}`);
     return response.data;
   } catch (error) {
-    const axiosError = error as AxiosError<ErrorMessage>; //  에러 타입 명시
-
-    if (axiosError.response) {
-      throw new Error(axiosError.response.data.message);
-    } else {
-      throw new Error("서버를 연결할 수 없습니다. 인터넷 연결을 확인해주세요.");
-    }
+    return handleApiError(error);
   }
 };
 
@@ -44,12 +28,6 @@ export const putShop = async (shopId: string, body: ShopRequest): Promise<ShopRe
     const response = await instance.put<ShopResponse>(`/shops/${shopId}`, body);
     return response.data;
   } catch (error) {
-    const axiosError = error as AxiosError<ErrorMessage>; //  에러 타입 명시
-
-    if (axiosError.response) {
-      throw new Error(axiosError.response.data.message);
-    } else {
-      throw new Error("서버를 연결할 수 없습니다. 인터넷 연결을 확인해주세요.");
-    }
+    return handleApiError(error);
   }
 };
